simulator build working
