+++ .vscode/launch.json
@@ -1,0 +1,13 @@
+{
+  "version": "0.2.0",
+  "configurations": [
+    {
+      "name": "C/C++ Runner: Debug Session",
+      "type": "lldb",
+      "request": "launch",
+      "args": [],
+      "cwd": "/Users/andrewyz/Code/Tec-App/TECApp",
+      "program": "/Users/andrewyz/Code/Tec-App/TECApp/build/Debug/outDebug"
+    }
+  ]
+}
