--- TECApp/App.js
@@ -1,42 +1,0 @@
-import { StyleSheet } from 'react-native'
-import React, {useEffect} from "react"
-import { NavigationContainer } from '@react-navigation/native'
-import { createNativeStackNavigator } from '@react-navigation/native-stack'
-import { Home } from './src/pages/Home'
-import { useFonts } from 'expo-font';
-import * as SplashScreen from 'expo-splash-screen';
-
-SplashScreen.preventAutoHideAsync();
-
-export default function App() {
-  const Stack = createNativeStackNavigator()
-  
-  const [loaded, error] = useFonts({
-    'Brix Sans': require('./assets/fonts/BrixSans.otf'),
-  });
-
-  useEffect(() => {
-    if (loaded || error) {
-      SplashScreen.hideAsync();
-    }
-  }, [loaded, error]);
-
-  if (!loaded && !error) {
-    return null;
-  }
-
-  return (
-    <NavigationContainer>
-    <Stack.Navigator
-      screenOptions={{
-        headerShown: false,
-        contentStyle: { backgroundColor: '#1C2B47' },
-      }}
-    >
-      <Stack.Screen name="Home" component={Home} />
-    </Stack.Navigator>
-  </NavigationContainer>
-  )
-}
-
-
