+++ TECApp/app/SVGs/DistributeRenewablesIcons/BiomassIcon.tsx
