+++ TECApp/app/SVGs/DistributeRenewablesIcons/GeothermalIcon.tsx
