+++ TECApp/app/SVGs/DistributeRenewablesIcons/HydroIcon.tsx
