+++ TECApp/app/SVGs/DistributeRenewablesIcons/NuclearIcon.tsx
