+++ TECApp/app/SVGs/DistributeRenewablesIcons/SolarIcon.tsx
