+++ TECApp/app/SVGs/DistributeRenewablesIcons/ToolTipIcon.tsx
