+++ TECApp/app/SVGs/DistributeRenewablesIcons/WindIcon.tsx
