+++ TECApp/app/SVGs/LeftArrow.tsx
