+++ TECApp/app/SVGs/RenewableIcon.tsx
