+++ TECApp/app/SVGs/RightArrow.tsx
