+++ TECApp/app/SVGs/TemperatureIcon.tsx
