+++ TECApp/app/components/DataVisualizations/BAUComparison.tsx
@@ -113,7 +113,7 @@ export const BAUComparison = ({
 
 const styles = StyleSheet.create({
   header: {
-    fontFamily: 'Brix Sans',
+    // fontFamily: 'Brix Sans',
     fontSize: 24,
     marginBottom: 8,
   },
