+++ TECApp/app/components/DataVisualizations/CarbonBudget.tsx
@@ -308,7 +308,7 @@ export const CarbonBudget = () => {
 
 const styles = StyleSheet.create({
   header: {
-    fontFamily: 'Brix Sans',
+    // fontFamily: 'Brix Sans',
     fontSize: 24,
     marginBottom: 8,
   },
