+++ TECApp/app/components/DataVisualizations/LineGraph.tsx
