+++ TECApp/app/components/DataVisualizations/TechnologyComparison.tsx
@@ -135,7 +135,7 @@ export const TechnologyComparison = ({ data }: TechnologyComparisonProps) => {
 
 const styles = StyleSheet.create({
   header: {
-    fontFamily: 'Brix Sans',
+    // fontFamily: 'Brix Sans',
     fontSize: 24,
     marginBottom: 8,
   },
