+++ TECApp/app/components/DistributeRenewables.tsx
