+++ TECApp/app/index.tsx
@@ -1,0 +1,11 @@
+import React from "react"
+import { Redirect } from "expo-router"
+
+
+export default function App() {
+  return (
+      <Redirect href="/pages/Home" />
+  )
+}
+
+
