+++ TECApp/app/util/Calculations.tsx
