+++ TECApp/app/util/ValueDictionaries.tsx
