+++ TECApp/package.json
@@ -1,52 +1,54 @@
 {
   "name": "awesomeproject",
   "version": "1.0.0",
-  "main": "node_modules/expo/AppEntry.js",
+  "main": "expo-router/entry",
   "type": "module",
   "scripts": {
     "start": "expo start --tunnel",
-    "android": "expo start --android",
-    "ios": "expo start --ios",
+    "android": "expo run:android",
+    "ios": "expo run:ios",
     "web": "expo start --web",
     "lint-fix": "(pnpm eslint src || true) && pnpm exec prettier src --write"
   },
   "dependencies": {
     "@babel/runtime": "^7.24.5",
-    "@expo/metro-runtime": "~3.1.3",
+    "@expo/metro-runtime": "~3.2.3",
     "@expo/ngrok": "^4.1.3",
     "@gorhom/bottom-sheet": "^4.6.3",
     "@miblanchard/react-native-slider": "^2.6.0",
     "@openspacelabs/react-native-zoomable-view": "^2.1.6",
     "@react-native-community/slider": "^4.5.2",
-    "@react-native-picker/picker": "^2.7.7",
+    "@react-native-picker/picker": "2.7.5",
     "@react-navigation/native": "^6.1.17",
     "@react-navigation/native-stack": "^6.9.26",
     "d3": "^7.9.0",
-    "expo": "^51.0.18",
+    "expo": "~51.0.28",
     "expo-checkbox": "~3.0.0",
-    "expo-font": "~12.0.7",
-    "expo-router": "^3.5.17",
-    "expo-splash-screen": "~0.27.5",
-    "expo-status-bar": "~1.11.1",
+    "expo-status-bar": "~1.12.1",
     "fast-xml-parser": "^4.4.1",
     "firebase": "^10.12.0",
     "react": "18.2.0",
     "react-dom": "18.2.0",
-    "react-native": "0.73.6",
+    "react-native": "0.74.5",
     "react-native-gesture-handler": "~2.16.1",
     "react-native-reanimated": "~3.10.1",
-    "react-native-svg": "^14.1.0",
+    "react-native-svg": "15.2.0",
     "react-native-svg-pan-zoom": "^0.1.2",
     "react-native-web": "^0.19.11",
     "react-router-dom": "^6.22.3",
-    "typescript": "^5.4.5"
+    "typescript": "~5.3.3",
+    "expo-router": "~3.5.23",
+    "react-native-safe-area-context": "4.10.5",
+    "react-native-screens": "3.31.1",
+    "expo-linking": "~6.3.1",
+    "expo-constants": "~16.0.2"
   },
   "devDependencies": {
     "@babel/core": "^7.24.5",
     "@eslint/js": "^9.2.0",
     "@types/d": "^1.0.4",
     "@types/d3": "^7.4.3",
-    "@types/react": "^18.3.1",
+    "@types/react": "~18.2.79",
     "eslint": "^8.57.0",
     "eslint-config-prettier": "^9.1.0",
     "eslint-plugin-react": "^7.34.1",
